Add cookies to return to previously visited weather forecast
Cookies are reintroduced mainly because firefox mobile works differently for PWAs than chrome.
Firefox mobile needs the start url to be set, which doesn't allow to create a PWA for the forecast of a specific city.
Now the PWA should return to the latest city a user requested a forecast for.